+++ rds/utility.py
@@ -10,24 +10,33 @@ import sys
 import json
 
 
-def get_response(api_call):
+def get_response(api_call, message=""):
     if sys.version_info > (3, 0):
         import urllib.request
-        return urllib.request.urlopen(api_call)
+
+        try:
+            return urllib.request.urlopen(api_call)
+        except urllib.error.HTTPError as e:
+            print(e)
+            print(message)
+            sys.exit()
     else:
         import urllib
-        return urllib.urlopen(api_call)
+
+        try:
+            return urllib.urlopen(api_call)
+        except urllib.HTTPError as e:
+            print(e)
+            print(message)
+            sys.exit()
 
 
 def check_valid(api_call, message, is_json=False):
     try:
-        response = get_response(api_call)
+        response = get_response(api_call, message=message)
         if is_json:
             return json.load(response)
-    except AttributeError as e:
-        print(e)
-        sys.exit
     except Exception as e:
-        print(message)
         print(e)
-        sys.exit
+        print(message)
+        sys.exit()